requirements.txt
# Core dependencies
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4

# Blockchain/Ethereum interaction
etherscan-python==2.1.0
web3==6.20.1

# Visualization
plotly==5.22.0
matplotlib==3.9.1
seaborn==0.13.2
jinja2==3.1.4

# Jupyter notebook support
jupyter==1.0.0
ipykernel==6.29.5
nbformat==5.10.4

# HTTP requests and API interaction
requests==2.32.3
urllib3==2.2.2
pycoingecko==3.1.0

# Data processing
python-dateutil==2.9.0
pytz==2024.1

# Development tools
ipython==8.26.0
black==24.4.2
pylint==3.2.6

# Optional but useful
tqdm==4.66.4  # Progress bars
tabulate==0.9.0  # Pretty tables